Fixed issue #45, RenamingDestinationFileDuringTransfer used workDir given by user and an exception was thrown
