+++ Frends.SFTP.DownloadFiles/Frends.SFTP.DownloadFiles.Tests/TransferTests.cs
@@ -184,227 +184,6 @@ class TransferTests : DownloadFilesTestBase
     }
 
     [Test]
-    public void DownloadFiles_TestAppendToExistingFile()
-    {
-        Directory.CreateDirectory(_destWorkDir);
-        File.Copy(Path.Combine(_workDir, _source.FileName), Path.Combine(_destWorkDir, _source.FileName));
-        var file1 = new FileInfo(Path.Combine(_workDir, _source.FileName));
-        Helpers.UploadTestFiles(new List<string> { Path.Combine(_workDir, "SFTPDownloadTestFile2.txt") }, _source.Directory);
-
-        var destination = new Destination
-        {
-            Directory = _destWorkDir,
-            FileName = _source.FileName,
-            Action = DestinationAction.Append,
-            FileNameEncoding = FileEncoding.UTF8,
-            EnableBomForFileName = true
-        };
-
-        var source = new Source
-        {
-            Directory = "/upload/Upload",
-            FileName = "SFTPDownloadTestFile2.txt",
-            Action = SourceAction.Error,
-            Operation = SourceOperation.Nothing
-        };
-
-        var result = SFTP.DownloadFiles(source, destination, _connection, _options, _info, new CancellationToken());
-        Assert.IsTrue(result.Success);
-        var file2 = new FileInfo(Path.Combine(_destWorkDir, _source.FileName));
-        Assert.AreNotEqual(file1.Length, file2.Length);
-    }
-
-    [Test]
-    public void DownloadFiles_SourceOperationNothingWithRenamingDisable()
-    {
-        Helpers.UploadTestFiles(new List<string> { Path.Combine(_workDir, _source.FileName) }, _source.Directory);
-
-        var options = new Options
-        {
-            ThrowErrorOnFail = false,
-            RenameSourceFileBeforeTransfer = false,
-            RenameDestinationFileDuringTransfer = false,
-            CreateDestinationDirectories = true,
-            PreserveLastModified = false,
-            OperationLog = true
-        };
-        var result = SFTP.DownloadFiles(_source, _destination, _connection, options, _info, new CancellationToken());
-        Assert.IsTrue(result.Success);
-
-        Assert.IsTrue(Helpers.SourceFileExists(_source.Directory + "/" + _source.FileName));
-    }
-
-    [Test]
-    public void DownloadFiles_TestSourceOperationMove()
-    {
-        var to = "uploaded";
-        Helpers.UploadTestFiles(new List<string> { Path.Combine(_workDir, _source.FileName) }, _source.Directory, to);
-        to = "upload/Upload/" + to;
-        var source = new Source
-        {
-            Directory = "upload/Upload",
-            FileName = "SFTPDownloadTestFile.txt",
-            Action = SourceAction.Error,
-            Operation = SourceOperation.Move,
-            DirectoryToMoveAfterTransfer = to
-        };
-
-        var result = SFTP.DownloadFiles(source, _destination, _connection, _options, _info, new CancellationToken());
-        Assert.IsTrue(result.Success);
-        Assert.AreEqual(1, result.SuccessfulTransferCount);
-
-        Assert.IsTrue(Helpers.SourceFileExists(source.DirectoryToMoveAfterTransfer + "/" + source.FileName));
-    }
-
-    [Test]
-    public void DownloadFiles_TestSourceOperationRename()
-    {
-        Helpers.UploadTestFiles(new List<string> { Path.Combine(_workDir, _source.FileName) }, _source.Directory);
-
-        var source = new Source
-        {
-            Directory = "upload/Upload",
-            FileName = "SFTPDownloadTestFile.txt",
-            Action = SourceAction.Error,
-            Operation = SourceOperation.Rename,
-            FileNameAfterTransfer = "uploaded_%SourceFileName%.txt"
-        };
-
-        var result = SFTP.DownloadFiles(source, _destination, _connection, _options, _info, new CancellationToken());
-        Assert.IsTrue(result.Success);
-        Assert.AreEqual(1, result.SuccessfulTransferCount);
-
-        Assert.IsTrue(Helpers.SourceFileExists(source.Directory + "/uploaded_" + source.FileName));
-    }
-
-    [Test]
-    public void DownloadFiles_TestSourceOperationMoveWithRenameFilesDuringTransfer()
-    {
-        var to = "uploaded";
-        Helpers.UploadTestFiles(new List<string> { Path.Combine(_workDir, _source.FileName) }, _source.Directory, to);
-        to = "upload/Upload/" + to;
-
-        var options = new Options
-        {
-            ThrowErrorOnFail = false,
-            RenameSourceFileBeforeTransfer = false,
-            RenameDestinationFileDuringTransfer = false,
-            CreateDestinationDirectories = true,
-            PreserveLastModified = false,
-            OperationLog = true
-        };
-
-        var source = new Source
-        {
-            Directory = "upload/Upload",
-            FileName = "SFTPDownloadTestFile.txt",
-            Action = SourceAction.Error,
-            Operation = SourceOperation.Move,
-            DirectoryToMoveAfterTransfer = to
-        };
-
-        var result = SFTP.DownloadFiles(source, _destination, _connection, options, _info, new CancellationToken());
-        Assert.IsTrue(result.Success);
-        Assert.AreEqual(1, result.SuccessfulTransferCount);
-
-        Assert.IsTrue(Helpers.SourceFileExists(source.DirectoryToMoveAfterTransfer + "/" + source.FileName));
-    }
-
-    [Test]
-    public void DownloadFiles_TestSourceOperationRenameWithRenameFilesDuringTransfer()
-    {
-        Helpers.UploadTestFiles(new List<string> { Path.Combine(_workDir, _source.FileName) }, _source.Directory);
-
-        var options = new Options
-        {
-            ThrowErrorOnFail = false,
-            RenameSourceFileBeforeTransfer = false,
-            RenameDestinationFileDuringTransfer = false,
-            CreateDestinationDirectories = true,
-            PreserveLastModified = false,
-            OperationLog = true
-        };
-
-        var source = new Source
-        {
-            Directory = "upload/Upload",
-            FileName = "SFTPDownloadTestFile.txt",
-            Action = SourceAction.Error,
-            Operation = SourceOperation.Rename,
-            FileNameAfterTransfer = "uploaded_%SourceFileName%.txt"
-        };
-
-        var result = SFTP.DownloadFiles(source, _destination, _connection, options, _info, new CancellationToken());
-        Assert.IsTrue(result.Success);
-        Assert.AreEqual(1, result.SuccessfulTransferCount);
-
-        Assert.IsTrue(Helpers.SourceFileExists(source.Directory + "/uploaded_" + source.FileName));
-    }
-
-    [Test]
-    public void DownloadFiles_TestSourceOperationMoveWithRenameFilesDuringTransferWithRenameSourceAndDestinationFilesEnabled()
-    {
-        var to = "uploaded";
-        Helpers.UploadTestFiles(new List<string> { Path.Combine(_workDir, _source.FileName) }, _source.Directory, to);
-        to = "upload/Upload/" + to;
-
-        var options = new Options
-        {
-            ThrowErrorOnFail = false,
-            RenameSourceFileBeforeTransfer = true,
-            RenameDestinationFileDuringTransfer = true,
-            CreateDestinationDirectories = true,
-            PreserveLastModified = false,
-            OperationLog = true
-        };
-
-        var source = new Source
-        {
-            Directory = "upload/Upload",
-            FileName = "SFTPDownloadTestFile.txt",
-            Action = SourceAction.Error,
-            Operation = SourceOperation.Move,
-            DirectoryToMoveAfterTransfer = to
-        };
-
-        var result = SFTP.DownloadFiles(source, _destination, _connection, options, _info, new CancellationToken());
-        Assert.IsTrue(result.Success);
-        Assert.AreEqual(1, result.SuccessfulTransferCount);
-
-        Assert.IsTrue(Helpers.SourceFileExists(source.DirectoryToMoveAfterTransfer + "/" + source.FileName));
-    }
-
-    [Test]
-    public void DownloadFiles_TestSourceOperationRenameWithRenameFilesDuringTransferWithRenameSourceAndDestinationFilesEnabled()
-    {
-        Helpers.UploadTestFiles(new List<string> { Path.Combine(_workDir, _source.FileName) }, _source.Directory);
-
-        var options = new Options
-        {
-            ThrowErrorOnFail = false,
-            RenameSourceFileBeforeTransfer = true,
-            RenameDestinationFileDuringTransfer = true,
-            CreateDestinationDirectories = true,
-            PreserveLastModified = false,
-            OperationLog = true
-        };
-
-        var source = new Source
-        {
-            Directory = "upload/Upload",
-            FileName = "SFTPDownloadTestFile.txt",
-            Action = SourceAction.Error,
-            Operation = SourceOperation.Rename,
-            FileNameAfterTransfer = "uploaded_%SourceFileName%.txt"
-        };
-
-        var result = SFTP.DownloadFiles(source, _destination, _connection, options, _info, new CancellationToken());
-        Assert.IsTrue(result.Success);
-        Assert.AreEqual(1, result.SuccessfulTransferCount);
-
-        Assert.IsTrue(Helpers.SourceFileExists(source.Directory + "/uploaded_" + source.FileName));
-    }
-    [Test]
     public void DownloadFiles_TestDownloadWithOverwrite()
     {
         Helpers.UploadTestFiles(new List<string> { Path.Combine(_workDir, _source.FileName) }, _source.Directory);
@@ -452,36 +231,6 @@ class TransferTests : DownloadFilesTestBase
         Assert.AreEqual(1, result.SuccessfulTransferCount);
 
         Assert.IsTrue(File.Exists(Path.Combine(destination.Directory, _source.FileName)));
-    }
-
-    [Test]
-    public void DownloadFiles_TestWithSourceMoveToNonExistingDirectoryShouldReturnUnsuccessfulTransfer()
-    {
-        Helpers.UploadTestFiles(new List<string> { Path.Combine(_workDir, _source.FileName) }, _source.Directory);
-        Directory.CreateDirectory(_destWorkDir);
-
-        var options = new Options
-        {
-            ThrowErrorOnFail = false,
-            RenameSourceFileBeforeTransfer = true,
-            RenameDestinationFileDuringTransfer = true,
-            CreateDestinationDirectories = true,
-            PreserveLastModified = true,
-            OperationLog = true
-        };
-
-        var source = new Source
-        {
-            Directory = _source.Directory,
-            FileName = _source.FileName,
-            Action = SourceAction.Error,
-            Operation = SourceOperation.Move,
-            DirectoryToMoveAfterTransfer = "/upload/test"
-        };
-
-        var result = SFTP.DownloadFiles(source, _destination, _connection, options, _info, new CancellationToken());
-        Assert.IsFalse(result.Success);
-        Assert.IsTrue(Helpers.SourceFileExists(_source.Directory + "/" + _source.FileName));
     }
 
     [Test]
